Remove mandatory component role specification
I left in most of the infrastructure for this -- at some point we'll
want to look deeper into specification of component roles, but it seems
silly to block translation (as was the case before this fix) in the end
to end scenario. Additionally, information to determine a device's role
is not currently in the DML, so this change would have to be percolated
up into that level of abstraction.